requirements.txt
customtkinter==5.2.2
pandas==2.0.3
openpyxl==3.1.2
pyzk==0.9
python-dotenv==1.0.0
Pillow==10.0.1
